Refactor/Checkstyle changes
