+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/DocmosisApplication.java
@@ -8,6 +8,7 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
 @EnableFeignClients(basePackages = {"uk.gov.hmcts.ethos.replacement"
     }
 )
+@SuppressWarnings("HideUtilityClassConstructor") // Spring needs a constructor, this is not a utility class
 public class DocmosisApplication {
 
     public static void main(String[] args) {
