+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/domain/repository/VenueRepository.java
@@ -2,8 +2,8 @@ package uk.gov.hmcts.ethos.replacement.docmosis.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import uk.gov.hmcts.ethos.replacement.docmosis.domain.referencedata.Venue;
 import uk.gov.hmcts.ecm.common.model.helper.TribunalOffice;
+import uk.gov.hmcts.ethos.replacement.docmosis.domain.referencedata.Venue;
 
 import java.util.List;
