+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/service/casetransfer/CaseTransferEventParams.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.ethos.replacement.docmosis.service;
+package uk.gov.hmcts.ethos.replacement.docmosis.service.casetransfer;
 
 import lombok.Builder;
 import lombok.Getter;
