+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/service/referencedata/CourtWorkerService.java
@@ -1,8 +1,8 @@
 package uk.gov.hmcts.ethos.replacement.docmosis.service.referencedata;
 
 import uk.gov.hmcts.ecm.common.model.bulk.types.DynamicValueType;
-import uk.gov.hmcts.ethos.replacement.docmosis.domain.referencedata.CourtWorkerType;
 import uk.gov.hmcts.ecm.common.model.helper.TribunalOffice;
+import uk.gov.hmcts.ethos.replacement.docmosis.domain.referencedata.CourtWorkerType;
 
 import java.util.List;
